Module value_router::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

CloseProgramAuthorityContext

Generated client accounts for CloseProgramAuthorityContext.

CreateRelayData

Generated client accounts for CreateRelayData.

InitializeContext

Generated client accounts for InitializeContext.

PostBridgeData

Generated client accounts for PostBridgeData.

PostSwapData

Generated client accounts for PostSwapData.

ReclaimContext

Generated client accounts for ReclaimContext.

RelayInstruction

Generated client accounts for RelayInstruction.

RelayNoSwapInstruction

Generated client accounts for RelayNoSwapInstruction.

SetAdminContext

Generated client accounts for SetAdminContext.

SetValueRouterContext

Generated client accounts for SetValueRouterContext.

SwapAndBridgeInstruction

Generated client accounts for SwapAndBridgeInstruction.

SwapAndBridgeShareEventAccountsInstruction

Generated client accounts for SwapAndBridgeShareEventAccountsInstruction.